requirements.txt
numpy>=1.24.1
opencv_python>=4.7.0
Pillow>=9.4.0
torch>=1.13.1
torchvision>=0.14.1
POT
tensordict~=0.5.0
--extra-index-url https://download.pytorch.org/whl/cu116
